refactor: add character utility functions for full-to-half conversion and hyphen detection
